# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Inventario API",
    "description": "REST API for Inventario — a multi-tenant, white-label asset management platform for sports federations, municipalities, schools, and clubs. Open-source under EUPL-1.2.",
    "version": "0.1.0",
    "contact": {
      "name": "Inventario · LTK Solutions",
      "url": "https://inventario.sportup.sk",
      "email": "inventario@sportup.sk"
    },
    "license": {
      "name": "EUPL-1.2",
      "url": "https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Microsoft Entra ID access token (v2.0). Multi-tenant: the JWT `tid` claim resolves the Organisation; the `oid` claim resolves the User. See apps/api/README.md for the device code flow."
      },
      "deploymentSecret": {
        "type": "http",
        "scheme": "bearer",
        "description": "Zdieľané tajomstvo pre servisné endpointy pod /v1/system — nie používateľský token. Migrácie a indexy používajú MIGRATIONS_SECRET (volá GitHub Actions po produkčnom deployi), retencia CRON_SECRET (volá Vercel Cron). Bez nastavenej premennej endpoint vracia 503."
      }
    },
...
  },
  "paths": {
...
    "/v1/public/organisations/login-context": {
      "get": {
        "operationId": "getV1PublicOrganisationsLoginContext",
        "summary": "Verejný login-context organizácie podľa slugu alebo vlastnej domény (ADR-0035)",
        "tags": [
          "Public"
        ],
        "description": "Bez autentifikácie. Vracia len neškodné dáta (branding + povolené spôsoby prihlásenia) potrebné na to, aby /login mohol zobraziť správne tlačidlá pre danú organizáciu. Presne jeden z query parametrov (slug/domain) musí byť zadaný.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 40
            },
            "in": "query",
            "name": "slug",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 253
            },
            "in": "query",
            "name": "domain",
            "required": false
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "slug": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "logoUrl": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "uri"
                    },
                    "brandColors": {
                      "type": [
                        "object",
                        "null"
                      ],
                      "properties": {
                        "primary": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "pattern": "^#[0-9a-fA-F]{6}$",
                          "description": "Hex farba #RRGGBB"
                        },
                        "primaryFg": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "pattern": "^#[0-9a-fA-F]{6}$",
                          "description": "Hex farba #RRGGBB"
                        },
                        "accent": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "pattern": "^#[0-9a-fA-F]{6}$",
                          "description": "Hex farba #RRGGBB"
                        },
                        "accentFg": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "pattern": "^#[0-9a-fA-F]{6}$",
                          "description": "Hex farba #RRGGBB"
                        }
                      },
                      "required": [
                        "primary",
                        "primaryFg",
                        "accent",
                        "accentFg"
                      ],
                      "additionalProperties": false
                    },
                    "allowedAuthProviders": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "hasEntraRestriction": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "slug",
                    "displayName",
                    "logoUrl",
                    "brandColors",
                    "allowedAuthProviders",
                    "hasEntraRestriction"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/public/scan/{publicToken}": {
      "get": {
        "operationId": "getV1PublicScanByPublicToken",
        "summary": "Verejny lost & found lookup po QR tokene (ADR-0021)",
        "tags": [
          "Public"
        ],
        "description": "Bez autentifikacie. Vracia minimalne informacie o najdenom majetku. 404 ak tenant nema zapnuty publicAssetLookup ALEBO token neexistuje (no oracle).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "publicToken",
            "required": true
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "organisationName": {
                      "type": "string"
                    },
                    "organisationLogoUrl": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "uri"
                    },
                    "foundContact": {
                      "type": [
                        "object",
                        "null"
                      ],
                      "properties": {
                        "email": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "phone": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "message": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      },
                      "required": [
                        "email",
                        "phone",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "organisationName",
                    "organisationLogoUrl",
                    "foundContact"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/system/indexes/ensure": {
      "post": {
        "operationId": "postV1SystemIndexesEnsure",
        "summary": "Vytvorenie MongoDB indexov po deployi",
        "tags": [
          "System"
        ],
        "description": "Volá GitHub Actions po úspešnom produkčnom deployi. V produkcii sa indexy pri boote nevytvárajú — detaily v lib/ensure-indexes.ts. Idempotentné.",
        "security": [
          {
            "deploymentSecret": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    },
    {
      "url": "https://api.inventario.sportup.sk",
      "description": "Production (planned Q3 2026)"
    }
  ],
...
}